patch1

+++ worldgen/terrain/source/cpu/elements/voronoi_rocks.cpp
@@ -18,7 +18,7 @@ extern "C" {
         int n_auxiliaries = 2;
         if (auxs == NULL) n_auxiliaries = 0;
         #pragma omp parallel for
-        for (size_t idx = 0; idx < size; idx++) {
+        for (int idx = 0; idx < size; idx++) {
             voronoi_rocks(
                 positions[idx], sdfs + idx, auxs + n_auxiliaries * idx, meta_param, second_meta_param,
                 d_i_params, d_f_params, second_d_i_params, second_d_f_params, third_d_i_params, third_d_f_params
