+++ worldgen/terrain/source/cpu/surfaces/chunkyrock.cpp
@@ -19,7 +19,7 @@ extern "C" {
         float3_nonbuiltin *offsets
     ) {
         #pragma omp parallel for
-        for (size_t idx = 0; idx < size; idx++) {
+        for (int idx = 0; idx < size; idx++) {
             geo_rocks(
                 positions[idx], normals[idx], f_params, f3_params, f4_params,
                 offsets + idx
