+++ worldgen/terrain/source/cpu/surfaces/mountain.cpp
@@ -18,7 +18,7 @@ extern "C" {
         float3_nonbuiltin *offsets
     ) {
         #pragma omp parallel for
-        for (size_t idx = 0; idx < size; idx++) {
+        for (int idx = 0; idx < size; idx++) {
             geo_MOUNTAIN(
                 positions[idx], normals[idx], f_params, f3_params, 
                 offsets + idx
