+++ worldgen/terrain/source/cpu/surfaces/mud.cpp
@@ -17,7 +17,7 @@ extern "C" {
         float3_nonbuiltin *offsets
     ) {
         #pragma omp parallel for
-        for (size_t idx = 0; idx < size; idx++) {
+        for (int idx = 0; idx < size; idx++) {
             geo_mud(
                 positions[idx], normals[idx], f_params, 
                 offsets + idx
